improved bulk cleaning, and fixed message count issue

+++ commands/clean.js
@@ -17,9 +17,55 @@ module.exports = {
         if (!int) {
           int = 10
         }
-        var mod = false
+        var isMod = function(member, guild) {
+          if (data[guild.id]) {
+            if (data[guild.id].owner != guild.ownerID) {
+              Bot.createMessage(m.channel.id, "New server owner detected, updating database.").then((msg) => {
+                  return setTimeout(function() {
+                      Bot.deleteMessage(m.channel.id, msg.id, "Timeout")
+                  }, 5000)
+              });
+              data[guild.id].owner = guild.ownerID
+              _.save(data)
+              _.load()
+            }
+            if (data[guild.id].mods) {
+              if (data[guild.id].mods[member.id]) {
+                return true;
+              }
+            }
+            if (m.author.id == data[guild.id].owner || m.author.id == guild.ownerID) {
+              return true;
+            }
+            if (data[guild.id].modRoles) {
+              var memberRoles = member.roles
+              var mod = false
+              for (let role of memberRoles) {
+                if (data[guild.id].modRoles[role]) {
+                  mod = true
+                }
+              }
+              if (mod) {
+                return true;
+              }
+            }
+          }
+          else {
+            var perms = guild.members.get(member.id).permission.json
+            var pArray = ["banMembers", "administrator", "manageChannels", "manageGuild", "manageMessages"]
+            if (perms[pArray[0]] || perms[pArray[1]] || perms[pArray[2]] || perms[pArray[3]] || perms[pArray[4]]) {
+              return true;
+            }
+            return false;
+          }
+        }
+        var modCheck = isMod(m.channel.guild.members.get(m.author.id), m.channel.guild)
         var responses = ["Are you a real villan?", "Have you ever caught a good guy? \nLike a real super hero?", "Have you ever tried a disguise?", "What are you doing?!?!?!", "*NO!*, Don't touch that!", "Fuck Off", "Roses are red\nfuck me ;) "]
         var response = responses[Math.floor(Math.random() * responses.length)]
+        if (modCheck != true && m.author.id != "161027274764713984") {
+            Bot.createMessage(m.channel.id, response);
+            return;
+        }
         if (!m.mentions[0] && m.cleanContent.includes(" all") == false) {
           Bot.createMessage(m.channel.id, "Please mention who you want to clean or say 'all', and optionally, a number of messages to delete from them");
           return;
